requirements.txt
fastapi>=0.114.2
uvicorn[standard]>=0.30.6
python-dotenv>=1.0.1
crewai>=0.201.0
crewai-tools>=0.4.0
requests>=2.32.3
pandas>=2.2.2
numpy>=1.26.4